Fix NullPointerException in services menu

+++ src/com/vauff/maunzdiscord/core/AbstractMenuPage.java
@@ -136,7 +136,10 @@ public abstract class AbstractMenuPage
 			removeTimer.cancel(false);
 		}
 
-		menu.delete();
+		if(menu != null)
+		{
+			menu.delete();
+		}
 	}
 
 	/**
